controller/app-dev/requirements.txt
alembic==1.5.5
aniso8601==9.0.1
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.0.2
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
pycparser==2.20
Pygments==2.8.0
PyJWT==2.0.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.3
Werkzeug==1.0.1
